Change a cast

+++ utils/update_countries.py
@@ -105,7 +105,7 @@ def get_countries_test():
     next(reader)
     for i in reader:
         if i[18] and i[4] and i[12] and i[20]:
-            percent = int(i[4])/float(i[12])*100
+            percent = float(i[4])/float(i[12])*100
             data[i[0]]['test_efectivity'].append(percent)
             data[i[0]]['total_tests_per_million'].append(float(i[12])*1000000/float(i[20]))
             data[i[0]]['population'].append(int(float(i[20])))
